requirements.txt
eventlet==0.25.1
Flask==2.0.2
numpy==1.22.2
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
pytest==7.1.2
pytest-mock==3.7.0
python-socketio==4.4.0
websocket-client==1.2.3
websockets==8.1
mypy
imutils